fix: emitSelected doesn't support multiple subscribers
Fixes an issue where the emitter returned by emitSelected would only
notify the first subscriber of changes. This happened because the Ref
that tracked the previous derived value was being shared between all the
subscribers. So, when the first subscription was processesed, the ref
would get updated, and then all subsequent subscriptions would read the
value, determine nothing had changed, and return early. This moves the
creation of the Ref inside the callback to makeEmitter, so that each new
subscriber gets its own Ref to track.