add runAndEmitStoreT
Adds a new entrypoint API function that allows the caller to obtain an
emitter from the store in addition to a component. This can be used to
react to store changes outside of the component.

An example application for this is persisting values from the store to
local storage or some other persistence mechanism. Without this
capability, one must perform these tasks inside their Halogen component's
`handleAction` when they receive new Input. This couples the component
with the persistence layer. Even if the concrete persistence mechanism is
injected, the component will still be aware that persistence is required.

With this concern handled outside the component, a "plug-in" style of
architecture is enabled, allowing the driver to opt-in or out of
persistence as desired. For example, a test driver that hosts a
component does not need to mock or configure any peristence layer to run
the component, but a production driver is still free to do so as
required.