Upgrade package set, add runAndEmitStoreT, selectEmitter, fork instances (#15)
* Upgrade package set to 0.14.7-20220303

* Add purescript-language-server to nix shell

* fix: emitSelected doesn't support multiple subscribers

Fixes an issue where the emitter returned by emitSelected would only
notify the first subscriber of changes. This happened because the Ref
that tracked the previous derived value was being shared between all the
subscribers. So, when the first subscription was processesed, the ref
would get updated, and then all subsequent subscriptions would read the
value, determine nothing had changed, and return early. This moves the
creation of the Ref inside the callback to makeEmitter, so that each new
subscriber gets its own Ref to track.

* add runAndEmitStoreT

Adds a new entrypoint API function that allows the caller to obtain an
emitter from the store in addition to a component. This can be used to
react to store changes outside of the component.

An example application for this is persisting values from the store to
local storage or some other persistence mechanism. Without this
capability, one must perform these tasks inside their Halogen component's
`handleAction` when they receive new Input. This couples the component
with the persistence layer. Even if the concrete persistence mechanism is
injected, the component will still be aware that persistence is required.

With this concern handled outside the component, a "plug-in" style of
architecture is enabled, allowing the driver to opt-in or out of
persistence as desired. For example, a test driver that hosts a
component does not need to mock or configure any peristence layer to run
the component, but a production driver is still free to do so as
required.

* add selectEmitter

Moved the logic of applying a selector to an emitter from
Halogen.Store.Monad to Halogen.Store.Select in a new function
`selectEmitter`. This allows the logic of applying a selector to an
emitter to be reused outside of `selectEmitted`, for example, when
running an application via `runAndEmitStoreT`.

* add fork instances for StoreT

- Adds `MonadFork` instance
- Adds `MonadKill` instance
- Adds `MonadBracket` instance

* Apply suggestions from code review

Thanks for the review @thomashoneyman

Co-authored-by: Thomas Honeyman <hello@thomashoneyman.com>

* Update src/Halogen/Store/Monad.purs

Co-authored-by: Thomas Honeyman <hello@thomashoneyman.com>